requirements.txt
numpy
matplotlib
scipy
opencv-python
pandas
google.generativeai
fpdf
flask
dlib
whisper
bertopic
sentence_transformers